refactor: relocate skill management and skill modules to `src/skills` directory.
