+++ src/skills/lib/multi-agent.js
