+++ src/skills/lib/skill-architect.js
