+++ src/skills/lib/skill-manager.js
