+++ src/skills/lib/tool-explorer.js
